Move entities to `entity` crate
