+++ Cargo.toml
@@ -35,3 +35,5 @@ sea-orm = { version = "0.7.1", features = [
 serde = "1.0.137"
 serde_json = "1.0.81"
 sha2 = "0.10.2"
+
+entity = { path = "entity" }
