+++ entity/Cargo.toml
@@ -6,3 +6,9 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+sea-orm = { version = "0.7.1", features = [
+  "sqlx-postgres",
+  "runtime-actix-rustls",
+  "macros",
+  "debug-print",
+] }
