+++ migration/Cargo.toml
@@ -11,6 +11,8 @@ path = "src/lib.rs"
 [dependencies]
 async-std = { version = "^1", features = ["attributes", "tokio1"] }
 
+entity = { path = "../entity" }
+
 [dependencies.sea-orm-migration]
 version = "^0.9.0"
 features = [
